Add api support
This is very WIP but this starts to add the ability for this demo app to
call a ZeeMee rails api to get basic user info.

Next steps are to wire this up further than just "auto fill form". We
have access to user images now but wiring that in this demo app is more
work.

Related PR on the rails side to support this
https://github.com/zeemee/zeemee/pull/3993